Building media uploaded.

+++ Telesto Basement App/CustomerDetailInfoObject.h
@@ -25,6 +25,6 @@
 @property (strong,nonatomic) NSString *customerNotes;
 //@property (strong,nonatomic) UIImage *customerProfileImage;
 @property (strong,nonatomic) NSDictionary *customerOtherImageDic;
-@property (assign,nonatomic) long latitude;
-@property (assign,nonatomic) long longitude;
+@property (assign,nonatomic) float latitude;
+@property (assign,nonatomic) float longitude;
 @end
